# Conflicts:
#	pharmacy_base_server/src/main/resources/application.properties
#	pharmacy_common/src/main/java/com/kgc/entity/CgrkOrder.java
#	pharmacy_finance_server/src/main/resources/application.properties
#	pharmacy_gateway/src/main/resources/application.yml
#	pharmacy_inventory_view/src/router/index.js
#	pharmacy_procurement_server/src/main/resources/application.properties
#	pharmacy_sale_server/src/main/resources/application.properties
#	pharmacy_system_server/src/main/resources/application.properties
#	pharmacy_warehouse_server/src/main/java/com/kgc/feign/PublicOMedicineService.java
#	pharmacy_warehouse_server/src/main/java/com/kgc/feign/SysUserRoleFeign.java
#	pharmacy_warehouse_server/src/main/resources/application.properties

+++ pharmacy_warehouse_server/src/main/java/com/kgc/service/impl/CgrkOrderServiceImpl.java
@@ -87,7 +87,7 @@ public class CgrkOrderServiceImpl extends ServiceImpl<CgrkOrderMapper, CgrkOrder
             referencCount = referencCount.add(multiply); // 使用BigDecimal的add方法进行精确加法计算
         }
         cgrqOrder.setCount(count);
-        cgrqOrder.setMedicineprice(referencCount.doubleValue());
+        cgrqOrder.setMedicineprice(referencCount);
         if (cgrqOrder.getOrderStatus()>0){
             cgrqOrder.setOrderStatus(cgrqOrder.getOrderStatus());
         }else {
